backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.0
authlib==1.2.1
itsdangerous==2.1.2

selenium==4.15.2
# webdriver-manager==4.0.1